Organize project creating packages
