+++ .gitignore
@@ -48,3 +48,4 @@ dist/**
 *.xlf
 package.nls.*.json
 l10n/
+env/**
